# Perform the zoom in/out canvas rendering with Cairo