This still needs some more testing, but has some good performance improvements in the benchmarks

## Current
|             Method |       Mean |     Error |     StdDev | Allocated |
|------------------- |-----------:|----------:|-----------:|----------:|
|     RenderOneToOne |   4.322 ms | 0.0942 ms |  0.2734 ms |     254 B |
| RenderManyOneToOne |  23.601 ms | 0.6252 ms |  1.8237 ms |     850 B |
|       RenderZoomIn |  38.219 ms | 0.7541 ms |  1.8640 ms |  28,399 B |
|   RenderManyZoomIn | 231.494 ms | 5.5985 ms | 16.3310 ms |  29,184 B |
|      RenderZoomOut |  11.879 ms | 0.2343 ms |  0.3434 ms |     288 B |
|  RenderManyZoomOut | 108.522 ms | 2.2333 ms |  6.4079 ms |   1,011 B |


## PR
|             Method |       Mean |     Error |    StdDev | Allocated |
|------------------- |-----------:|----------:|----------:|----------:|
|     RenderOneToOne |   4.263 ms | 0.0734 ms | 0.0651 ms |     206 B |
| RenderManyOneToOne |  21.894 ms | 0.4364 ms | 0.6664 ms |     448 B |
|       RenderZoomIn |  16.880 ms | 0.2621 ms | 0.2451 ms |     226 B |
|   RenderManyZoomIn | 119.798 ms | 1.6916 ms | 1.5823 ms |     579 B |
|      RenderZoomOut |   3.944 ms | 0.0697 ms | 0.0953 ms |     203 B |
|  RenderManyZoomOut |  31.815 ms | 0.6140 ms | 1.3607 ms |     467 B |