@cameronwhite 2022-01-19
Also attaching the original benchmarks stats for my machine from before the previous PR

|             Method |       Mean |      Error |     StdDev | Allocated |
|------------------- |-----------:|-----------:|-----------:|----------:|
|     RenderOneToOne |   4.065 ms |  0.0808 ms |  0.2266 ms |     246 B |
| RenderManyOneToOne |  21.894 ms |  0.4324 ms |  0.7105 ms |     842 B |
|       RenderZoomIn |  65.521 ms |  1.2960 ms |  2.5583 ms |  28,430 B |
|   RenderManyZoomIn | 596.972 ms | 11.3052 ms | 15.4747 ms |  29,936 B |
|      RenderZoomOut |  11.486 ms |  0.2278 ms |  0.3547 ms |     277 B |
|  RenderManyZoomOut | 106.560 ms |  2.1096 ms |  2.7430 ms |   1,219 B |