@jpobst 2022-01-20
Very nice, definitely feels faster to use!